op-update/requirements.txt
flask
minio
psycopg2-binary
requests
datetime